# Optimize performance for renderPageRectToImageData method