Hi @jichang , could you take a look? This might be faster, from `478ms` to `178ms`, or from `1000ms` to `500ms`, depends on the size.