@nguyenyou 2024-11-18
You might be interested in the approach used in this https://github.com/ajrcarey/pdfium-render/issues/32